Update and Delete functionlaity improved with logged in user

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModal");
+const User = require("../models/userModal");
 
 // @desc Get Goals
 // @route GET /api/goals/
@@ -47,6 +48,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error(`Goal not found`);
   }
 
+  //  Check  for user
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error(`User not found`);
+  }
+
+  //  Making sure the login user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error(`User not authorized`);
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -61,6 +76,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error(`Goal not found`);
+  }
+  //  Check  for user
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error(`User not found`);
+  }
+
+  //  Making sure the login user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error(`User not authorized`);
   }
   await Goal.findOneAndDelete(req.params.id);
   // await goal.remove();
